перенес функции поиска фильмов в App
