+++ src/components/App/App.js
@@ -4,7 +4,6 @@ import React, { useState, useEffect } from 'react'
 import { Context } from '../../context/CurrentUserContext';
 import { useNavigate } from 'react-router-dom';
 
-
 import Header from '../Header/Header';
 import Promo from '../Promo/Promo';
 import AboutProject from '../AboutProject/AboutProject';
@@ -21,6 +20,7 @@ import Error from '../Error/Error';
 import BurgerMenu from '../BurgerMenu/BurgerMenu';
 import Main from '../Main/Main';
 import apiMain from '../../utils/MainApi';
+import api from '../../utils/MoviesApi';
 
 function App() {
   const navigate = useNavigate();
@@ -30,6 +30,17 @@ function App() {
   const [currentUser, setUserInfo] = useState({});
   const [burger, openBurger] = useState(false);
   const [nameProfile, setNameProfile] = useState('');
+  //movies const
+  const [moreButton, setMoreButton] = useState(false);
+  const [films, getFilms] = useState([]);
+  const [searchFilms, getSearchFilms] = useState([]);
+  const [next, setNext] = useState(0);
+  const [saveFilms, getSaveFilms] = useState([]);
+  const [isComponentSaveFilms, getIsComponentSaveFilms] = useState(false);
+
+  const filmsPerPageForBigWidth = 3;
+  const filmsPerPageForSmalWidth = 2;
+  let arrayForHoldingFilms = [];
 
   useEffect(() => {
     const jwt = localStorage.getItem('jwt');
@@ -64,6 +75,67 @@ function App() {
     }
   }, [loggedIn])
 
+  // Movies
+  useEffect(() => {
+    api.getFilm()
+      .then((res) => {
+        let allFilms = res;
+        getFilms(allFilms);
+      }).catch((err) => {
+        alert(err.message);
+      });
+  }, [])
+
+  // Save movies
+  useEffect(() => {
+    const jwt = localStorage.getItem('jwt');
+    apiMain.getSavedFilm(jwt)
+      .then((res) => {
+        let allFilms = res;
+        getSaveFilms(allFilms);
+      }).catch((err) => {
+        alert(err);
+      });
+  }, []);
+
+  //Movies
+  function getSerchFilm(word) {
+    if (word === '') {
+      return
+    }
+    localStorage.setItem('searchWord', word);
+    if (window.matchMedia("(max-width: 600px)").matches) {
+      showTheNumberOfFilms(0, 5)
+    }
+    else if (window.matchMedia("(max-width: 800px)").matches) {
+      showTheNumberOfFilms(0, 8)
+    }
+    else if (window.matchMedia("(max-width: 1400px)").matches) {
+      showTheNumberOfFilms(0, 12)
+    }
+  }
+
+  function showTheNumberOfFilms(start, end) {
+    let word = localStorage.getItem('searchWord');
+    let findFilms = !isComponentSaveFilms ? films.filter(element => element.nameRU.match(word)) : saveFilms.filter(element => element.nameRU.match(word))
+    const slicedFilms = findFilms.slice(start, end);
+    arrayForHoldingFilms = [...arrayForHoldingFilms, ...slicedFilms];
+    getSearchFilms(arrayForHoldingFilms);
+    setNext(end);
+    setMoreButton(arrayForHoldingFilms.length < findFilms.length);
+  }
+
+  const handleShowMorePosts = () => {
+    if (window.matchMedia("(max-width: 800px)").matches) {
+      showTheNumberOfFilms(0, next + filmsPerPageForSmalWidth);
+      setNext(next + filmsPerPageForSmalWidth);
+    } else if (window.matchMedia("(max-width: 1400px)").matches) {
+      showTheNumberOfFilms(0, next + filmsPerPageForBigWidth);
+      setNext(next + filmsPerPageForBigWidth);
+    }
+  };
+
+  //Burger
   function handleOpenBurger() {
     openBurger(true)
   }
@@ -80,15 +152,9 @@ function App() {
       .then((res) => {
         localStorage.setItem('jwt', res.token);
         navigate('/signin');
-        // setPopupAuthorization(true)
-        // setSuccessful({ image: Successful, message: 'Вы успешно зарегистрировались!' })
       })
       .catch((res) => {
         alert(res);
-        // setPopupAuthorization(true)
-        // setSuccessful({
-        //   image: Mistake, message: 'Что-то пошло не так! Попробуйте ещё раз.'
-        // })
       })
   }
 
@@ -104,10 +170,6 @@ function App() {
       })
       .catch((err) => {
         alert(err);
-        // setPopupAuthorization(true)
-        // setSuccessful({
-        //   image: Mistake, message: 'Что-то пошло не так! Попробуйте ещё раз.'
-        // })
       });
   }
 
@@ -145,28 +207,43 @@ function App() {
           </>} />
           <Route path='/movies' element={<>
             <Header
+              getSearchFilms={getSearchFilms}
               handleOpenBurger={handleOpenBurger}
               loggedIn={loggedIn} />
             <Main>
               <Movies
+                getIsComponentSaveFilms={getIsComponentSaveFilms}
+                handleShowMorePosts={handleShowMorePosts}
+                getSerchFilm={getSerchFilm}
+                searchFilms={searchFilms}
+                moreButton={moreButton}
                 isOn={value}
                 handleToggle={() => setValue(!value)}
               />
             </Main>
             <Footer />
           </>} />
-          <Route path='/saved-movies' element={<>
-            <Header
-              handleOpenBurger={handleOpenBurger}
-              loggedIn={loggedIn} />
-            <Main>
-              <SavedMovies
-                isOn={value}
-                handleToggle={() => setValue(!value)}
-              />
-            </Main>
-            <Footer />
-          </>} />
+          <Route
+            path='/saved-movies'
+            element={<>
+              <Header
+                getSearchFilms={getSearchFilms}
+                handleOpenBurger={handleOpenBurger}
+                loggedIn={loggedIn} />
+              <Main>
+                <SavedMovies
+                  getIsComponentSaveFilms={getIsComponentSaveFilms}
+                  getSerchFilm={getSerchFilm}
+                  searchFilms={searchFilms}
+                  moreButton={moreButton}
+                  handleShowMorePosts={handleShowMorePosts}
+                  isOn={value}
+                  getSearchFilms={getSearchFilms}
+                  handleToggle={() => setValue(!value)}
+                />
+              </Main>
+              <Footer />
+            </>} />
           <Route path='/profile' element={<>
             <Header
               loggedIn={loggedIn} />
