+++ src/components/Header/Header.js
@@ -2,7 +2,11 @@ import './Header.css';
 import logo from '../../images/logo.svg'
 import { Link } from 'react-router-dom';
 
-function Header({ loggedIn, handleOpenBurger }) {
+function Header({ loggedIn, handleOpenBurger, getSearchFilms }) {
+
+  function clearStateSearch() {
+    getSearchFilms([]);
+  }
 
   return (
     loggedIn ?
@@ -11,8 +15,8 @@ function Header({ loggedIn, handleOpenBurger }) {
           <img className="header__logo" src={logo} alt='логотип' />
         </Link>
         <div className='header__navigation'>
-          <Link className='header__navigation-button' to='/movies'>Фильмы</Link>
-          <Link className='header__navigation-button' to='/saved-movies'>Сохранённые фильмы</Link>
+          <Link onClick={clearStateSearch} className='header__navigation-button' to='/movies'>Фильмы</Link>
+          <Link onClick={clearStateSearch} className='header__navigation-button' to='/saved-movies'>Сохранённые фильмы</Link>
         </div>
         <Link className='header__profile' to='/profile'>Аккаунт</Link>
         <button onClick={handleOpenBurger} className='header__burger'></button>
